refactoring of various app engine operations
