+++ google-cloud-tools-plugin/resources/messages/CloudToolsBundle.properties
@@ -57,7 +57,7 @@ appengine.config.version.error=Enter a version ID.
 appengine.flex.deployment.cost.warning={0}There is no free quota for App Engine flexible environment deployments. Please visit {1}GCP Pricing{2} for pricing information.{3}
 appengine.flex.user.specified.deploymentsource.name=Filesystem JAR or WAR file
 appengine.flex.user.specified.deploymentsource.name.with.filename=Filesystem JAR or WAR file - {0}
-appengine.action.credential.not.found=We were unable to find a credential for your cloud project's user. Are you sure {0} is logged in?
+appengine.action.credential.not.found=We were unable to find a credential for your cloud project's user. Are you sure you are logged in?
 appengine.deployment.error.creating.staging.directory=There was an unexpected error creating the staging directory
 appengine.generate.config.button=Generate...
 appengine.appyaml.location.label=app.yaml path\:
