+++ google-cloud-tools-plugin/src/com/google/cloud/tools/intellij/appengine/cloud/AppEngineDeploymentConfigurator.java
@@ -156,6 +156,7 @@ class AppEngineDeploymentConfigurator extends
         project,
         source,
         new CloudSdkAppEngineHelper(
+            project,
             new File(server.getConfiguration().getCloudSdkHomePath()),
             getEnvironment())
     );
